# Add support for smbus2 i2c_rdwr to avoid unnecessary 32byte chunking