@thijstriemstra 2019-04-16
@rm-hull can you adjust the coverage treshold so it doesn't report broken builds for small changes? Maybe something like 1 or 1.5% would be better?