@rm-hull 2019-05-10
v1.9.0 released to pypi with this improvement incorporated